[ANDROAPP-4048] feature config package refactor
