+++ app/src/main/AndroidManifest.xml
@@ -158,7 +158,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".usescases.featureconfig.FeatureConfigView" />
+        <activity android:name=".usescases.featureconfig.ui.FeatureConfigView" />
 
     </application>
