+++ app/src/main/assets/paperwork.json
@@ -1,1 +1,1 @@
-{"buildTime":"2021-06-16 07:53","gitSha":"6a996a4aa"}
+{"buildTime":"2021-06-16 08:57","gitSha":"8914e6eb4"}
