+++ app/src/main/java/org/dhis2/data/user/UserComponent.java
@@ -32,8 +32,8 @@ import org.dhis2.usescases.eventsWithoutRegistration.eventCapture.EventCaptureCo
 import org.dhis2.usescases.eventsWithoutRegistration.eventCapture.EventCaptureModule;
 import org.dhis2.usescases.eventsWithoutRegistration.eventInitial.EventInitialComponent;
 import org.dhis2.usescases.eventsWithoutRegistration.eventInitial.EventInitialModule;
-import org.dhis2.usescases.featureconfig.FeatureConfigComponent;
-import org.dhis2.usescases.featureconfig.FeatureConfigModule;
+import org.dhis2.usescases.featureconfig.di.FeatureConfigComponent;
+import org.dhis2.usescases.featureconfig.di.FeatureConfigModule;
 import org.dhis2.usescases.main.MainComponent;
 import org.dhis2.usescases.main.MainModule;
 import org.dhis2.usescases.main.program.ProgramComponent;
