+++ app/src/main/java/org/dhis2/usescases/development/DevelopmentActivity.java
@@ -5,7 +5,6 @@ import android.content.res.Resources;
 import android.os.Build;
 import android.os.Bundle;
 import android.util.DisplayMetrics;
-import android.view.View;
 import android.widget.Toast;
 
 import androidx.annotation.Nullable;
@@ -17,7 +16,7 @@ import com.google.gson.reflect.TypeToken;
 import org.dhis2.App;
 import org.dhis2.R;
 import org.dhis2.databinding.DevelopmentActivityBinding;
-import org.dhis2.usescases.featureconfig.FeatureConfigView;
+import org.dhis2.usescases.featureconfig.ui.FeatureConfigView;
 import org.dhis2.usescases.general.ActivityGlobalAbstract;
 import org.dhis2.usescases.main.MainActivity;
 import org.dhis2.utils.customviews.BreakTheGlassBottomDialog;
