--- app/src/main/java/org/dhis2/usescases/featureconfig/Feature.kt
@@ -1,8 +1,0 @@
-package org.dhis2.usescases.featureconfig
-
-/**
- * Enum Issue name
- * description user explanation
- * ex: ANDROAPP_2557("Local analytics")
- */
-enum class Feature(val description: String)
