+++ app/src/main/java/org/dhis2/usescases/featureconfig/di/FeatureConfigComponent.kt
@@ -1,7 +1,8 @@
-package org.dhis2.usescases.featureconfig
+package org.dhis2.usescases.featureconfig.di
 
 import dagger.Subcomponent
 import org.dhis2.data.dagger.PerActivity
+import org.dhis2.usescases.featureconfig.ui.FeatureConfigView
 
 @PerActivity
 @Subcomponent(
