+++ app/src/main/java/org/dhis2/usescases/featureconfig/di/FeatureConfigModule.kt
@@ -1,9 +1,11 @@
-package org.dhis2.usescases.featureconfig
+package org.dhis2.usescases.featureconfig.di
 
 import dagger.Module
 import dagger.Provides
 import org.dhis2.data.dagger.PerActivity
 import org.dhis2.data.prefs.PreferenceProvider
+import org.dhis2.usescases.featureconfig.data.FeatureConfigRepository
+import org.dhis2.usescases.featureconfig.data.FeatureConfigRepositoryImpl
 
 @Module
 class FeatureConfigModule {
