+++ app/src/main/java/org/dhis2/usescases/featureconfig/model/Feature.kt
@@ -1,0 +1,3 @@
+package org.dhis2.usescases.featureconfig.model
+
+enum class Feature(val description: String)
