+++ app/src/main/java/org/dhis2/usescases/featureconfig/model/FeatureState.kt
@@ -1,3 +1,3 @@
-package org.dhis2.usescases.featureconfig
+package org.dhis2.usescases.featureconfig.model
 
 data class FeatureState(val feature: Feature, val enable: Boolean)
